feat: added Grants Treasury + Tests and got to 100% coverage on treasury

+++ test/glow-token/Handler.sol
@@ -5,17 +5,18 @@ import "forge-std/StdUtils.sol";
 import "forge-std/StdCheats.sol";
 import "forge-std/console.sol";
 import {IGlow} from "../../src/interfaces/IGlow.sol";
-contract Handler is StdUtils,StdCheats {
+
+contract Handler is StdUtils, StdCheats {
     Glow public glow;
     bool boundVars = true;
 
     constructor(address _glow) public {
         glow = Glow(_glow);
-        deal(address(this),1e18 ether);
+        deal(address(this), 1e18 ether);
     }
 
     function stake(uint256 amount) public {
-        if(boundVars){
+        if (boundVars) {
             amount = bound(amount, 0, glow.balanceOf(address(this)));
         }
         //log sender
@@ -24,26 +25,26 @@ contract Handler is StdUtils,StdCheats {
     }
 
     function unstake(uint256 amount) public {
-        if(boundVars){
+        if (boundVars) {
             amount = bound(amount, 0, glow.numStaked(address(this)));
         }
-        console.log("SENDER = " , msg.sender);
+        console.log("SENDER = ", msg.sender);
 
         glow.unstake(amount);
     }
 
-    function claimUnstakedTokens(uint amount) public {
-        if(boundVars){
+    function claimUnstakedTokens(uint256 amount) public {
+        if (boundVars) {
             amount = bound(amount, 0, getUnstakedBalance());
         }
         glow.claimUnstakedTokens(amount);
     }
-    
+
     function getUnstakedBalance() public view returns (uint256) {
         IGlow.UnstakedPosition[] memory unstakedPositions = glow.unstakedPositionsOf(address(this));
-        uint counter;
-        unchecked{
-            for(uint i; i<unstakedPositions.length; i++){
+        uint256 counter;
+        unchecked {
+            for (uint256 i; i < unstakedPositions.length; i++) {
                 counter += unstakedPositions[i].amount;
             }
         }
